feat: slug to direct url
