+++ next.config.js
@@ -51,6 +51,12 @@ let nextConfig = {
         permanent: true,
         basePath: false,
       },
+      {
+        source: "/strategies/:slug/manage",
+        destination: "/strategies/:slug",
+        permanent: true,
+        basePath: false,
+      },
     ]
   },
 }
